ヘッダーの簡素化とナビメニューの追加 (#178)
* feat: Add Navigation

* feat: Simplify Header

* feat: Apply on Layout

* fix: Add transition and Adjust

* style: Add margin-right
