+++ src/components/header.tsx
@@ -1,29 +1,14 @@
-import { Button } from "./button";
 import type { FC } from "react";
 import Link from "next/link";
 import styles from "../scss/components/header.module.scss";
 
-export interface LinksProps {
-  links: readonly Readonly<{
-    name: string;
-    url: string;
-  }>[];
-}
-
-const Links: FC<LinksProps> = ({ links }) => (
-  <nav className={styles.buttonWrapper}>
-    {links.map(({ name, url }) => (
-      <Button key={name}>
-        <Link href={url}>
-          <a>{name}</a>
-        </Link>
-      </Button>
-    ))}
-  </nav>
-);
-
-export const Header: FC<LinksProps> = (props) => (
+export const Header: FC = () => (
   <header className={styles.headerLayout}>
-    <Links {...props} />
+    <Link href="/">
+      <a className={styles.hero}>
+        <img src="/android-chrome-512x512.png"></img>
+        Approvers
+      </a>
+    </Link>
   </header>
 );
