+++ src/scss/components/navigation.module.scss
@@ -1,0 +1,46 @@
+@use "../details.variable";
+@use "../color.variable";
+@use "../media.variable";
+
+.buttonWrapper {
+  @include details.flex-style();
+
+  flex-flow: column;
+  align-items: flex-start;
+
+  & > div {
+    width: 100%;
+    padding: 0 1rem;
+    margin: 1rem 0;
+  }
+}
+
+.navButton {
+  @include color.text(sub, front);
+  @include color.sub(dark, back);
+
+  position: fixed;
+  bottom: 0.5rem;
+  left: 0.5rem;
+  z-index: 20;
+  width: 4rem;
+  height: 4rem;
+  font-size: 2rem;
+  border-radius: 50%;
+}
+
+.navWrapper {
+  @include color.main(normal, border);
+  @include color.sub(light, back);
+
+  position: fixed;
+  right: 0.5rem;
+  bottom: 0.5rem;
+  z-index: 10;
+  border: thin solid;
+  transition: opacity 0.13s;
+}
+
+.navWrapper[data-showing="false"] {
+  opacity: 0;
+}
